# src/main/resources/assets/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Whiskly API",
        "description": "A better way to cook.",
        "version": "1.0.0",
        "contact": {
            "name": "Breland Miley",
            "email": "breland+api@whisk.ly",
            "url": "http://whisk.ly"
        }
    },
    "host": "api.whisk.ly",
    "schemes": [
        "https"
    ],
    "basePath": "/v1",
    "produces": [
        "application/json"
    ],
    "paths": {
        "/recipes/{id}": {
            "get": {
                "summary": "Single Recipe",
                "description": "Returns a single recipe for a given id.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the recipe to fetch",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    }
                ],
                "tags": [
                    "Recipes"
                ],
                "responses": {
                    "200": {
                        "description": "A recipe",
                        "schema": {
                            "$ref": "#/definitions/Recipe"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/recipes": {
            "get": {
                "summary": "Search all recipes",
                "description": "Returns a listing of all recipes, paged.",
                "parameters": [
                    {
                        "name": "pageNum",
                        "in": "query",
                        "description": "The page number to give back.",
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "The size of each page to give back.",
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "type": "string",
                        "description": "Searches all text fields."
                    }
                ],
                "tags": [
                    "Recipes",
                    "Search"
                ],
                "responses": {
                    "200": {
                        "description": "Successfuly ran a search and these matched.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Recipe"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "New Recipe",
                "description": "Adds a new recipe to the database, adding a unique id on the way in.",
                "parameters": [
                    {
                        "name": "Recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Recipe"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully added a new recipe.",
                        "schema": {
                            "$ref": "#/definitions/Recipe"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "tags": [
                    "Recipes"
                ]
            }
        }
    },
    "definitions": {
        "Recipe": {
            "properties": {
                "recipe_uuid": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Unique identifier representing a specific recipe."
                },
                "recipe_name": {
                    "type": "string",
                    "description": "Description of product."
                },
                "oven_fan": {
                    "$ref": "#/definitions/OvenFan"
                },
                "oven_time": {
                    "type": "integer",
                    "format": "int64"
                },
                "oven_temp": {
                    "$ref": "#/definitions/OvenTemp"
                },
                "yields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Yield"
                    }
                },
                "notes": {
                    "$ref": "#/definitions/Note"
                },
                "source_book": {
                    "$ref": "#/definitions/SourceBook"
                },
                "source_authors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Author"
                    }
                },
                "source_url": {
                    "type": "string"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Step"
                    }
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/IngredientAndAmount"
                    }
                },
                "prep_time": {
                    "type": "integer",
                    "format": "int64",
                    "description": "The time in minutes it takes to prep for a recipe."
                },
                "cooking_time": {
                    "type": "integer",
                    "format": "int64",
                    "description": "The time in minutes for this recipe to cook."
                },
                "added_at": {
                    "type": "integer",
                    "format": "int64",
                    "description": "The timestamp in milliseconds when this was created relative to the unix epoch."
                },
                "is_private": {
                    "type": "boolean"
                }
            }
        },
        "Error": {
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                },
                "fields": {
                    "type": "string"
                }
            }
        },
        "Author": {
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "HACCP": {
            "properties": {
                "critical_control_point": {
                    "type": "string"
                },
                "control_point": {
                    "type": "string"
                }
            }
        },
        "Note": {
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "note": {
                    "type": "string"
                }
            }
        },
        "Ingredient": {
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "name": {
                    "type": "string"
                },
                "usda_num": {
                    "type": "string"
                },
                "substitutions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Ingredient"
                    }
                },
                "note": {
                    "$ref": "#/definitions/Note"
                },
                "processingTags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ProcessingTag"
                    }
                }
            }
        },
        "UnitTag": {
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "tag": {
                    "type": "string"
                }
            }
        },
        "IngredientAndAmount": {
            "properties": {
                "id": {
                    "type": "string",
                    "foramt": "uuid"
                },
                "ingredient": {
                    "$ref": "#/definitions/Ingredient"
                },
                "value": {
                    "type": "number",
                    "format": "double"
                },
                "unit": {
                    "$ref": "#/definitions/UnitTag"
                }
            }
        },
        "OvenTemp": {
            "properties": {
                "value": {
                    "type": "integer",
                    "format": "int64"
                },
                "unit": {
                    "$ref": "#/definitions/TemperatureUnit"
                }
            }
        },
        "TemperatureUnit": {
            "properties": {
                "value": {
                    "type": "string",
                    "enum": [
                        "C",
                        "F"
                    ]
                }
            }
        },
        "OvenFan": {
            "properties": {
                "value": {
                    "type": "string",
                    "enum": [
                        "Off",
                        "Low",
                        "High"
                    ]
                }
            }
        },
        "ProcessingTag": {
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "tag": {
                    "type": "string"
                }
            }
        },
        "SourceBook": {
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "authors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Author"
                    }
                },
                "title": {
                    "type": "string"
                },
                "ISBN": {
                    "type": "string"
                },
                "notes": {
                    "$ref": "#/definitions/Note"
                }
            }
        },
        "Step": {
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "step_details": {
                    "type": "string"
                },
                "haccp": {
                    "$ref": "#/definitions/HACCP"
                },
                "notes": {
                    "$ref": "#/definitions/Note"
                },
                "order": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
        "Yield": {
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "serves": {
                    "type": "integer",
                    "format": "int64"
                },
                "unit": {
                    "$ref": "#/definitions/UnitTag"
                }
            }
        }
    }
}